Added contact form with EmailJS integration
The contact form sends the contents of its fields to the provided email. After submission, a pop-up is displayed with information about the successful or unsuccessful submission of the form. Unfortunately, it is not written in TypeScript, as there was a critical error which I was not able to resolve.